flattened file structure
