+++ access/datasets.py
@@ -7,8 +7,8 @@ import pandas as pd
 class datasets(object):
     _dir = 'chi_med_data'
 
-    _dir_path = os.path.join('./', _dir)
-    _abs_path = os.path.abspath(_dir_path)
+    _homedir = os.path.expanduser("~")
+    _dir_path = os.path.join(_homedir, _dir)
 
     _dwnld_data = {'chi_times' : 'https://drive.google.com/uc?authuser=0&id=1IcfJimPj4C5ZN5Xc-nvq_DModcCO6GY3&export=download',
                    'chi_euclidean' : 'https://drive.google.com/uc?authuser=0&id=1qq5ZWOaq5uxJOu9QzsNCw5WhdATgIhzK&export=download',
@@ -51,7 +51,7 @@ class datasets(object):
             path = os.path.join(datasets._dir_path, datasets._datasets[key])
 
             if key in datasets._dwnld_data.keys() and not os.path.exists(path):
-                print('Downloading {key} to {path}...'.format(key = key, path = datasets._abs_path))
+                print('Downloading {key} to {path}...'.format(key = key, path = datasets._dir_path))
                 req = requests.get(datasets._dwnld_data[key])
                 file_path = os.path.join(datasets._dir_path, datasets._datasets[key])
