--- access/fca/__init__.py
@@ -1,4 +1,0 @@
-from .fca1 import *
-from .fca2 import *
-from .fca3 import *
-
