--- access/helpers/__init__.py
@@ -1,2 +1,0 @@
-from .helpers import *
-
