--- access/raam/__init__.py
@@ -1,2 +1,0 @@
-from .raam import *
-
