--- access/weights/__init__.py
@@ -1,1 +1,0 @@
-from .weights import *
